requirements.txt
python-fasthtml
google-cloud-speech
websockets # For potential future real-time audio streaming (optional for now)
python-dotenv
google-generativeai
google-genai
boto3
